fix: make body as a component
